requirements.txt
# List your project dependencies here
weaviate
weaviate-client
pymupdf
pymupdf4llm
python-dotenv
openai
streamlit
sentence_transformers
nltk
markitdown
python-dotenv